migration table type fixes

+++ src/core/database/migrations/20220115100043-create_order_products.js
@@ -13,7 +13,7 @@ module.exports = {
       'order_products',
       {
         id: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
           autoIncrement: true,
           unique: true,
@@ -21,12 +21,12 @@ module.exports = {
         },
 
         order_id: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
 
         product_id: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
 
@@ -41,12 +41,12 @@ module.exports = {
         },
 
         quantity: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
 
         price: {
-          type: Sequelize.BIGINT,
+          type: Sequelize.INTEGER,
           allowNull: false,
         },
